% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ghc_authenticate_gql}
\alias{ghc_authenticate_gql}
\title{Authenticate for GraphQL queries}
\usage{
ghc_authenticate_gql(token)
}
\arguments{
\item{token}{authentication token for github API (https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/)}
}
\value{
\link[ghql]{GraphqlClient} object with authentication headers
}
\description{
Helper function to authenticate the GitHub account for running GraphQL (gql) queries. This function is exported for users who want to run custom GraphQL queries against the GitHub server but does not need to be called directly for standard use of the ghctools package. Note that the validity of the authentication token will only be tested once a query (\link{ghc_run_gql}) is run.
}
\seealso{
Other GitHub query functions: \code{\link{ghc_run_gql}}
}
