% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{ghc_clone_repositories}
\alias{ghc_clone_repositories}
\title{Clone repositories}
\usage{
ghc_clone_repositories(repos, folder = ".", token = NULL, pull = TRUE)
}
\arguments{
\item{repos}{data frame with, at minimum, columns 'repository' (name) and 'url'}

\item{folder}{the target directory where to clone all the repositories to}

\item{token}{the authentication token (only required for private repositories if global credentials are not set)}

\item{pull}{if the repository already exists, whether to update it from the remote (see \code{\link{ghc_repos_git_pull}})}
}
\value{
the \code{repos} data frame (unaltered) for easy piping
}
\description{
Clone repositories
}
