% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{ghc_repos_get_github_information}
\alias{ghc_repos_get_github_information}
\title{Get information on GitHub Classroom from GitHub}
\usage{
ghc_repos_get_github_information(prefix, org, token, max = 100)
}
\arguments{
\item{prefix}{assignment prefix (i.e. repository prefix set for github classroom assignment)}

\item{org}{organization name (i.e. organization github user)}

\item{token}{authentication token for github API (https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/)}

\item{max}{maximum number of repositories to retrieve information for}
}
\value{
data frame with repository names, urls, last commits, last pull request
}
\description{
Get information on GitHub Classroom from GitHub
}
\seealso{
Other information functions: \code{\link{ghc_repos_get_local_information}}
}
