% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{ghc_repos_get_local_information}
\alias{ghc_repos_get_local_information}
\title{Get local information on GitHub Classroom repositories}
\usage{
ghc_repos_get_local_information(prefix, folder = ".")
}
\arguments{
\item{prefix}{repository prefix (typically this would be a github classroom assignment prefix)}

\item{folder}{the target directory where to look for matching git repositories (searches recursively in all subdirectories)}
}
\value{
paths to the found repositories (relative to the working directory unless param \code{folder} is an absolute path)
}
\description{
Get local information on GitHub Classroom repositories
}
\seealso{
Other information functions: \code{\link{ghc_repos_get_github_information}}
}
