% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{ghc_repos_git_add}
\alias{ghc_repos_git_add}
\title{Git Add}
\usage{
ghc_repos_git_add(prefix, filepath, folder = ".")
}
\arguments{
\item{prefix}{repository prefix (typically this would be a github classroom assignment prefix)}

\item{filepath}{new file to start tracking with git. The filepath is relative to the main repository directory within each repository.}

\item{folder}{the target directory where to look for matching git repositories (searches recursively in all subdirectories)}
}
\value{
returns the \code{prefix} invisibly to facilitate piping
}
\description{
Adds the specified file to the active branch for all repositories that match the \code{prefix}. Use \code{\link{ghc_repos_git_commit}} to commit added files, followed by \code{\link{ghc_repos_git_push}} to push them to the server. Note that this function will always add the specific filed and overwrite .gitignore rules if necessary. You can also use this function to tag an existing file to be included in the next commit (this happens automatically for all modified files if you use the \code{\link{ghc_repos_git_commit}} function to commit changes).
}
\seealso{
Other command functions: 
\code{\link{ghc_repos_copy_to}()},
\code{\link{ghc_repos_git_clone}()},
\code{\link{ghc_repos_git_commit}()},
\code{\link{ghc_repos_git_create_branch}()},
\code{\link{ghc_repos_git_delete_branch}()},
\code{\link{ghc_repos_git_discard_changes}()},
\code{\link{ghc_repos_git_pull}()},
\code{\link{ghc_repos_git_push}()},
\code{\link{ghc_repos_git_remove}()},
\code{\link{ghc_repos_git_status}()},
\code{\link{ghc_repos_git_switch_branch}()},
\code{\link{ghc_repos_run_git_command}()}
}
\concept{command functions}
