% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{ghc_repos_git_clone}
\alias{ghc_repos_git_clone}
\title{Clone assignment repositories}
\usage{
ghc_repos_git_clone(prefix, org, folder = ".", token = NULL, pull = TRUE)
}
\arguments{
\item{prefix}{assignment prefix (i.e. repository prefix set for github classroom assignment)}

\item{org}{organization name (i.e. organization github user)}

\item{folder}{the target directory where to clone all the repositories to}

\item{token}{authentication token for github API (https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/)}

\item{pull}{if the repository already exists, whether to update it from the remote (see \code{\link{ghc_repos_git_pull}})}
}
\value{
repositories data frame, see \link{ghc_repos_git_find_repositories} for details
}
\description{
Clone all students' repositories for a specific assignment. Token with private repo access privileges is required if the students' repositories are private. This function can also be used to update existing local repositories for an assignment (i.e. after initial cloning if the students have made updates in the mean time). If updates are not desired, set \code{pull=TRUE}.
}
\seealso{
Other command functions: 
\code{\link{ghc_repos_copy_to}()},
\code{\link{ghc_repos_git_add}()},
\code{\link{ghc_repos_git_commit}()},
\code{\link{ghc_repos_git_create_branch}()},
\code{\link{ghc_repos_git_delete_branch}()},
\code{\link{ghc_repos_git_discard_changes}()},
\code{\link{ghc_repos_git_pull}()},
\code{\link{ghc_repos_git_push}()},
\code{\link{ghc_repos_git_remove}()},
\code{\link{ghc_repos_git_status}()},
\code{\link{ghc_repos_git_switch_branch}()},
\code{\link{ghc_repos_run_git_command}()}
}
\concept{command functions}
