% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{ghc_repos_git_commit}
\alias{ghc_repos_git_commit}
\title{Git Commit}
\usage{
ghc_repos_git_commit(prefix, message, folder = ".")
}
\arguments{
\item{prefix}{repository prefix (typically this would be a github classroom assignment prefix)}

\item{folder}{the target directory where to look for matching git repositories (searches recursively in all subdirectories)}
}
\value{
returns the \code{prefix} invisibly to facilitate piping
}
\description{
Commits changes for all modified files in the active branch for all repositories that match the \code{prefix}. Use \code{\link{ghc_repos_git_add}} and \code{\link{ghc_repos_git_remove}} prior to \code{\link{ghc_repos_git_commit}} to include new files (not tracked by git yet) / remove tracked files in the commit. Use \code{\link{ghc_repos_git_push}} to push the committed changes back to the GitHub server.
}
\seealso{
Other command functions: 
\code{\link{ghc_repos_copy_to}()},
\code{\link{ghc_repos_git_add}()},
\code{\link{ghc_repos_git_clone}()},
\code{\link{ghc_repos_git_create_branch}()},
\code{\link{ghc_repos_git_delete_branch}()},
\code{\link{ghc_repos_git_discard_changes}()},
\code{\link{ghc_repos_git_pull}()},
\code{\link{ghc_repos_git_push}()},
\code{\link{ghc_repos_git_remove}()},
\code{\link{ghc_repos_git_status}()},
\code{\link{ghc_repos_git_switch_branch}()},
\code{\link{ghc_repos_run_git_command}()}
}
\concept{command functions}
