% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{ghc_repos_git_push}
\alias{ghc_repos_git_push}
\title{Git Push}
\usage{
ghc_repos_git_push(prefix, folder = ".")
}
\arguments{
\item{prefix}{repository prefix (typically this would be a github classroom assignment prefix)}

\item{folder}{the target directory where to look for matching git repositories (searches recursively in all subdirectories)}
}
\value{
returns the \code{prefix} invisibly to facilitate piping
}
\description{
Pushes committed changes in the active branch (see \code{\link{ghc_repos_git_commit}}) for all repositories that match the \code{prefix} to GitHub. Note that this requires the repositories to have been cloned with a valid access token, otherwise this operation will fail due to lack of permissions. GitHub will also not allow \code{git push} if there are unpulled changes on the server. Use \code{\link{ghc_repos_git_pull}} followed by \code{\link{ghc_repos_git_push}} in this case (unless there are merging issues).
}
\seealso{
Other command functions: 
\code{\link{ghc_repos_copy_to}()},
\code{\link{ghc_repos_git_add}()},
\code{\link{ghc_repos_git_clone}()},
\code{\link{ghc_repos_git_commit}()},
\code{\link{ghc_repos_git_create_branch}()},
\code{\link{ghc_repos_git_delete_branch}()},
\code{\link{ghc_repos_git_discard_changes}()},
\code{\link{ghc_repos_git_pull}()},
\code{\link{ghc_repos_git_remove}()},
\code{\link{ghc_repos_git_status}()},
\code{\link{ghc_repos_git_switch_branch}()},
\code{\link{ghc_repos_run_git_command}()}
}
\concept{command functions}
