% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{ghc_repos_git_switch_branch}
\alias{ghc_repos_git_switch_branch}
\title{Git Switch Branch}
\usage{
ghc_repos_git_switch_branch(prefix, branch = "master", folder = ".")
}
\arguments{
\item{prefix}{repository prefix (typically this would be a github classroom assignment prefix)}

\item{branch}{branch name (for commands that interact with branches)}

\item{folder}{the target directory where to look for matching git repositories (searches recursively in all subdirectories)}
}
\value{
returns the \code{prefix} invisibly to facilitate piping
}
\description{
Lets you switch to a different branch that already exists for all repositories that match the \code{prefix}. Note that the git command used for this is \code{git checkout}.
}
\seealso{
Other command functions: \code{\link{ghc_repos_copy_to}},
  \code{\link{ghc_repos_git_add}},
  \code{\link{ghc_repos_git_clone}},
  \code{\link{ghc_repos_git_commit}},
  \code{\link{ghc_repos_git_create_branch}},
  \code{\link{ghc_repos_git_delete_branch}},
  \code{\link{ghc_repos_git_discard_changes}},
  \code{\link{ghc_repos_git_pull}},
  \code{\link{ghc_repos_git_push}},
  \code{\link{ghc_repos_git_remove}},
  \code{\link{ghc_repos_git_status}},
  \code{\link{ghc_repos_run_git_command}}
}
