% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ghc_run_gql}
\alias{ghc_run_gql}
\title{Run a GrahQL query}
\usage{
ghc_run_gql(
  query,
  query_name = "query",
  gql_client = ghc_authenticate_gql(token),
  token = NULL
)
}
\arguments{
\item{query}{GraphQL query}

\item{gql_client}{a GraphQL client object - will be generated based on the token by default}

\item{token}{authentication token (only needed if \code{gql_client} is not provided)}
}
\value{
returns the $data of the query
}
\description{
Run a GraphQL query against the github graphql node. Stops if errors are encountered during query execution. This function is exported for users who want to run custom GraphQL queries against the GitHub server but does not need to be called directly for standard use of the ghctools package.
}
\seealso{
Other GitHub query functions: 
\code{\link{ghc_authenticate_gql}()}
}
\concept{GitHub query functions}
