% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ghctools-package}
\alias{ghctools}
\alias{ghctools-package}
\title{ghctools: Tools for GitHub Classroom}
\description{
Tools for using GitHubClassroom.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/ghctools}
  \item Report bugs at \url{https://github.com/kopflab/ghctools/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
